Update all libreadline easyconfig to link to ncurses
The full explanation for this can be found in issue #1315.

libreadline needs some terminal info. It can get this from libtermcap or
ncurses, which are (as far as I understand it) API but not ABI
compatible. On most (=all the ones I looked at) systems, libtermcap just
gives a redirect to libtinfo. This libraries is created by splitting a
part of the main ncurses library because most program only need this and
not full ncurses. In EB we don't do this and it's one big fat library.
Also, libreadline and ncurses don't mix. It's next to impossible to use
both of them in the same code.

The idea is that, the code that uses libreadline can choose which
termcap library to use, either libtermcap or ncurses. But as said
before, on most distro's the leads to the same library.

So, my proposed solution is: we explicitly link libreadline to ncurses.
This will definitely solve the problem in this PR but I'm not entirely
sure about potential fallout. It seems like most EB stuff already links
to ncurses when using libreadline. I'm still guessing about why this
issue does not pop up in RHEL (and derivates) but on Debian we do hit
it.

Relevant bugreport: https://bugzilla.redhat.com/show_bug.cgi?id=499837