Add missing dependency on XZ to libxml2.
The original problem is that when building libxml2 on a system with
liblzma installed by the distro, it picks up dependencies on lzma
routines. Some of the versions of Gromacs are linking against libxml2
statically, causing link errors agains lzma. And any other software that
does static linking against libxml2 will have the same problem.

There is a cyclic dependency between XZ, libxml2, and gettext.
Fix this by adding a dummy toolchain build of gettext without any
dependencies on anything.

XZ for the updated toolchains are then built with the
dummy-toolchain gettext as a build dependency.
And then normal toolchain builds of libxml2 (now with XZ dependency) and
gettext can be built against that XZ.

This was originally discussed in issue #3408.